[ADD] sale_management_crm: create rental orders from crm
When Both apps sale_management and sale_renting are installed, sale_renting
has the priority.

Create a rental quotation from opportunity when crm and sale_renting module are
installed.

Add rental quotation and order stat button in crm lead form view.

If one product in the order is rentable, the SO is a rental order. Otherwise,
a normal sale order is used.

Task: 2081209
Closes: #39297