[FIX] module: warn when module license is implicit
Module authors tend to forget specifying the license for their modules,
which means they default to LGPL-3. This is not much of an issue for
Community Edition modules, but it's a real problem for Enterprise
Edition modules, and third-party modules, where it could create an
apparent conflict with the expected license of the app/product.

Adding a warning without changing the behavior should help prevent
those omissions. It's better to specify the license explicitly even in
Community Edition modules.